+++ client/src/components/web-extension/Send/Send.scss
@@ -9,4 +9,11 @@
 			width: 100%;
 		}
 	}
+	.cd_we_confirm {
+		.cd_we_confirm_row {
+			.cd_we_confirm_address {
+				word-break: break-all;
+			}
+		}
+	}
 }
